requirements.txt
Django>=5.0
djangorestframework>=3.14
dj-database-url>=2.0
gunicorn>=21.2
psycopg2-binary>=2.9
whitenoise>=6.5
python-decouple>=3.8 